server/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.1
torchvision==0.16.1
transformers
deepspeed
rtmlib
onnxruntime-gpu
huggingface_hub
azure-cognitiveservices-speech>=1.37.0
openai>=1.30.0
websockets>=12.0
python-dotenv>=1.0.0
httpx>=0.27.0
azure-cosmos>=4.7.0
azure-communication-rooms>=1.1.0
azure-communication-identity>=1.5.0
azure-identity>=1.17.0
azure-keyvault-secrets>=4.8.0
opencensus-ext-azure>=1.1.0
pydantic>=2.0.0
azure-storage-blob>=12.19.0
responsibleai>=0.34.0
raiwidgets>=0.34.0
reportlab>=4.0.0
